🚚 Move Footer to pattern
