+++ src/components/Layout/index.js
@@ -3,14 +3,15 @@ import PropTypes from "prop-types"
 import styled, { ThemeProvider } from "styled-components"
 
 import Navigation from "../Navigation"
-import Footer from "../Footer/Footer"
 import { Box } from "../Box"
+
+import { Footer } from "../../patterns"
 
 import Global from "./global"
 import theme from "../../../theme"
 
 const MainContainer = styled(Box)`
-  padding: 40px ${props => props.theme.spacing["4"]};
+  padding: 40px ${(props) => props.theme.spacing["4"]};
 
   @media (min-width: 768px) {
     padding: 60px 0 32px 60px;
